% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadings2table.R
\name{loadings2table.matrix}
\alias{loadings2table.matrix}
\alias{loadings2table}
\alias{loadings2table.fa}
\alias{loadings2table.fa.ci}
\alias{loadings2table.loadings}
\title{Factor loading table}
\usage{
\method{loadings2table}{matrix}(M, hints = character(0), Vexpl = numeric(0))

loadings2table(x, ...)

\method{loadings2table}{fa}(FA, ...)

\method{loadings2table}{fa.ci}(FA, ...)

\method{loadings2table}{loadings}(L, ...)
}
\arguments{
\item{M}{numeric matrix, dimnames must be present}

\item{hints}{named character vector with item content}

\item{Vexpl}{numeric vector with proportion of explained variance}

\item{...}{the arguments passed to \code{loadings2table.matrix}:
hints or Vexpl}

\item{FA}{\code{fa} or \code{fa.ci} object from \code{psych} package}

\item{L}{\code{loadings} object from \code{psych} package}
}
\description{
Render the table with factor loadings as html-widget.
}
